Move route validator
