+++ internal/k8s/reconciler/validators/route_test.go
@@ -1,4 +1,4 @@
-package reconciler
+package validators
 
 import (
 	"context"
@@ -8,7 +8,6 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/hashicorp/consul-api-gateway/internal/k8s/service"
 	"github.com/hashicorp/consul-api-gateway/internal/k8s/service/mocks"
-	"github.com/hashicorp/go-hclog"
 	"github.com/stretchr/testify/require"
 	gw "sigs.k8s.io/gateway-api/apis/v1alpha2"
 )
@@ -21,12 +20,7 @@ func TestRouteValidate(t *testing.T) {
 	resolver := mocks.NewMockBackendResolver(ctrl)
 
 	validator := NewRouteValidator(resolver)
-	factory := NewFactory(FactoryConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-
-	route := factory.NewRoute(&gw.HTTPRoute{})
-	state, err := validator.Validate(context.Background(), route)
+	state, err := validator.Validate(context.Background(), &gw.HTTPRoute{})
 	require.NoError(t, err)
 	require.True(t, state.ResolutionErrors.Empty())
 
@@ -39,7 +33,7 @@ func TestRouteValidate(t *testing.T) {
 	}
 	resolver.EXPECT().Resolve(gomock.Any(), gomock.Any(), reference).Return(resolved, nil)
 
-	route = factory.NewRoute(&gw.HTTPRoute{
+	state, err = validator.Validate(context.Background(), &gw.HTTPRoute{
 		Spec: gw.HTTPRouteSpec{
 			Rules: []gw.HTTPRouteRule{{
 				BackendRefs: []gw.HTTPBackendRef{{
@@ -50,17 +44,36 @@ func TestRouteValidate(t *testing.T) {
 			}},
 		},
 	})
-	state, err = validator.Validate(context.Background(), route)
 	require.NoError(t, err)
 	require.True(t, state.ResolutionErrors.Empty())
 
 	expected := errors.New("expected")
 	resolver.EXPECT().Resolve(gomock.Any(), gomock.Any(), reference).Return(nil, expected)
-	_, err = validator.Validate(context.Background(), route)
+	_, err = validator.Validate(context.Background(), &gw.HTTPRoute{
+		Spec: gw.HTTPRouteSpec{
+			Rules: []gw.HTTPRouteRule{{
+				BackendRefs: []gw.HTTPBackendRef{{
+					BackendRef: gw.BackendRef{
+						BackendObjectReference: reference,
+					},
+				}},
+			}},
+		},
+	})
 	require.Equal(t, expected, err)
 
 	resolver.EXPECT().Resolve(gomock.Any(), gomock.Any(), reference).Return(nil, service.NewK8sResolutionError("error"))
-	state, err = validator.Validate(context.Background(), route)
+	state, err = validator.Validate(context.Background(), &gw.HTTPRoute{
+		Spec: gw.HTTPRouteSpec{
+			Rules: []gw.HTTPRouteRule{{
+				BackendRefs: []gw.HTTPBackendRef{{
+					BackendRef: gw.BackendRef{
+						BackendObjectReference: reference,
+					},
+				}},
+			}},
+		},
+	})
 	require.NoError(t, err)
 	require.False(t, state.ResolutionErrors.Empty())
 }
